+++ model/src/main/java/com/vaadin/addon/charts/model/ColumnOptions.java
@@ -82,19 +82,6 @@ public abstract class ColumnOptions extends AbstractPlotOptions {
     public abstract void setBorderWidth(Number borderWidth);
 
     /**
-     * @see #setColor(Color)
-     */
-    public abstract Color getColor();
-
-    /**
-     * The main color or the series. In line type series it applies to the line
-     * and the point markers unless otherwise specified. In bar type series it
-     * applies to the bars unless a color is specified per point. The default
-     * value is pulled from the <code>options.colors</code> array.
-     */
-    public abstract void setColor(Color color);
-
-    /**
      * @see #setColorByPoint(Boolean)
      */
     public abstract Boolean getColorByPoint();
