+++ model/src/main/java/com/vaadin/addon/charts/model/DataLabelsFunnel.java
@@ -1,17 +1,13 @@
 package com.vaadin.addon.charts.model;
 
 import com.vaadin.addon.charts.model.style.Color;
-import com.vaadin.addon.charts.model.style.Style;
 /**
  */
 public class DataLabelsFunnel extends AbstractDataLabels {
 
-	private Color backgroundColor;
-	private Color borderColor;
 	private Number borderRadius;
 	private Number borderWidth;
 	private String className;
-	private Color color;
 	private Color connectorColor;
 	private Number connectorPadding;
 	private Number connectorWidth;
@@ -28,7 +24,6 @@ public class DataLabelsFunnel extends AbstractDataLabels {
 	private Boolean shadow;
 	private Shape shape;
 	private Boolean softConnector;
-	private Style style;
 	private Boolean useHTML;
 	private VerticalAlign verticalAlign;
 	private Number x;
@@ -36,35 +31,6 @@ public class DataLabelsFunnel extends AbstractDataLabels {
 	private Number zIndex;
 
 	public DataLabelsFunnel() {
-	}
-
-	/**
-	 * @see #setBackgroundColor(Color)
-	 */
-	public Color getBackgroundColor() {
-		return backgroundColor;
-	}
-
-	/**
-	 * The background color or gradient for the data label. Defaults to
-	 * <code>undefined</code>.
-	 */
-	public void setBackgroundColor(Color backgroundColor) {
-		this.backgroundColor = backgroundColor;
-	}
-
-	/**
-	 * @see #setBorderColor(Color)
-	 */
-	public Color getBorderColor() {
-		return borderColor;
-	}
-
-	/**
-	 * The border color for the data label. Defaults to <code>undefined</code>.
-	 */
-	public void setBorderColor(Color borderColor) {
-		this.borderColor = borderColor;
 	}
 
 	/**
@@ -117,20 +83,6 @@ public class DataLabelsFunnel extends AbstractDataLabels {
 	 */
 	public void setClassName(String className) {
 		this.className = className;
-	}
-
-	/**
-	 * @see #setColor(Color)
-	 */
-	public Color getColor() {
-		return color;
-	}
-
-	/**
-	 * The text color for the data labels. Defaults to <code>null</code>.
-	 */
-	public void setColor(Color color) {
-		this.color = color;
 	}
 
 	/**
@@ -420,35 +372,6 @@ public class DataLabelsFunnel extends AbstractDataLabels {
 	 */
 	public void setSoftConnector(Boolean softConnector) {
 		this.softConnector = softConnector;
-	}
-
-	/**
-	 * @see #setStyle(Style)
-	 */
-	public Style getStyle() {
-		if (style == null) {
-			style = new Style();
-		}
-		return style;
-	}
-
-	/**
-	 * Styles for the label. The default <code>color</code> setting is
-	 * <code>"contrast"</code>, which is a pseudo color that Highcharts picks up
-	 * and applies the maximum contrast to the underlying point item, for
-	 * example the bar in a bar chart. The <code>textOutline</code> is a pseudo
-	 * property that applies an outline of the given width with the given color,
-	 * which by default is the maximum contrast to the text. So a bright text
-	 * color will result in a black text outline for maximum readability on a
-	 * mixed background. In some cases, especially with grayscale text, the text
-	 * outline doesn't work well, in which cases it can be disabled by setting
-	 * it to <code>"none"</code>.
-	 * <p>
-	 * Defaults to: {"color": "contrast", "fontSize": "11px", "fontWeight":
-	 * "bold", "textOutline": "1px contrast" }
-	 */
-	public void setStyle(Style style) {
-		this.style = style;
 	}
 
 	/**
