+++ model/src/main/java/com/vaadin/addon/charts/model/DataLabelsRange.java
@@ -1,7 +1,4 @@
 package com.vaadin.addon.charts.model;
-
-import com.vaadin.addon.charts.model.style.Color;
-import com.vaadin.addon.charts.model.style.Style;
 /**
  * Extended data labels for range series types. Range series data labels have no
  * <code>x</code> and <code>y</code> options. Instead, they have
@@ -13,12 +10,9 @@ public class DataLabelsRange extends AbstractDataLabels {
 
 	private HorizontalAlign align;
 	private Boolean allowOverlap;
-	private Color backgroundColor;
-	private Color borderColor;
 	private Number borderRadius;
 	private Number borderWidth;
 	private String className;
-	private Color color;
 	private Boolean crop;
 	private Boolean defer;
 	private Boolean enabled;
@@ -30,7 +24,6 @@ public class DataLabelsRange extends AbstractDataLabels {
 	private Number rotation;
 	private Boolean shadow;
 	private Shape shape;
-	private Style style;
 	private Boolean useHTML;
 	private VerticalAlign verticalAlign;
 	private Number xHigh;
@@ -83,35 +76,6 @@ public class DataLabelsRange extends AbstractDataLabels {
 	}
 
 	/**
-	 * @see #setBackgroundColor(Color)
-	 */
-	public Color getBackgroundColor() {
-		return backgroundColor;
-	}
-
-	/**
-	 * The background color or gradient for the data label. Defaults to
-	 * <code>undefined</code>.
-	 */
-	public void setBackgroundColor(Color backgroundColor) {
-		this.backgroundColor = backgroundColor;
-	}
-
-	/**
-	 * @see #setBorderColor(Color)
-	 */
-	public Color getBorderColor() {
-		return borderColor;
-	}
-
-	/**
-	 * The border color for the data label. Defaults to <code>undefined</code>.
-	 */
-	public void setBorderColor(Color borderColor) {
-		this.borderColor = borderColor;
-	}
-
-	/**
 	 * @see #setBorderRadius(Number)
 	 */
 	public Number getBorderRadius() {
@@ -161,20 +125,6 @@ public class DataLabelsRange extends AbstractDataLabels {
 	 */
 	public void setClassName(String className) {
 		this.className = className;
-	}
-
-	/**
-	 * @see #setColor(Color)
-	 */
-	public Color getColor() {
-		return color;
-	}
-
-	/**
-	 * The text color for the data labels. Defaults to <code>null</code>.
-	 */
-	public void setColor(Color color) {
-		this.color = color;
 	}
 
 	/**
@@ -363,35 +313,6 @@ public class DataLabelsRange extends AbstractDataLabels {
 	 */
 	public void setShape(Shape shape) {
 		this.shape = shape;
-	}
-
-	/**
-	 * @see #setStyle(Style)
-	 */
-	public Style getStyle() {
-		if (style == null) {
-			style = new Style();
-		}
-		return style;
-	}
-
-	/**
-	 * Styles for the label. The default <code>color</code> setting is
-	 * <code>"contrast"</code>, which is a pseudo color that Highcharts picks up
-	 * and applies the maximum contrast to the underlying point item, for
-	 * example the bar in a bar chart. The <code>textOutline</code> is a pseudo
-	 * property that applies an outline of the given width with the given color,
-	 * which by default is the maximum contrast to the text. So a bright text
-	 * color will result in a black text outline for maximum readability on a
-	 * mixed background. In some cases, especially with grayscale text, the text
-	 * outline doesn't work well, in which cases it can be disabled by setting
-	 * it to <code>"none"</code>.
-	 * <p>
-	 * Defaults to: {"color": "contrast", "fontSize": "11px", "fontWeight":
-	 * "bold", "textOutline": "1px contrast" }
-	 */
-	public void setStyle(Style style) {
-		this.style = style;
 	}
 
 	/**
