+++ model/src/main/java/com/vaadin/addon/charts/model/PlotOptionsErrorbar.java
@@ -1,8 +1,8 @@
 package com.vaadin.addon.charts.model;
 
-import com.vaadin.addon.charts.model.style.Color;
 import java.util.ArrayList;
 import java.util.Arrays;
+import com.vaadin.addon.charts.model.style.Color;
 import java.util.Date;
 import java.time.Instant;
 import com.vaadin.addon.charts.util.Util;
@@ -16,7 +16,6 @@ public class PlotOptionsErrorbar extends AbstractPlotOptions {
 	private Boolean allowPointSelect;
 	private Number animationLimit;
 	private String className;
-	private Color color;
 	private Boolean colorByPoint;
 	private Number colorIndex;
 	private ArrayList<Color> colors;
@@ -115,24 +114,6 @@ public class PlotOptionsErrorbar extends AbstractPlotOptions {
 	 */
 	public void setClassName(String className) {
 		this.className = className;
-	}
-
-	/**
-	 * @see #setColor(Color)
-	 */
-	public Color getColor() {
-		return color;
-	}
-
-	/**
-	 * The main color of the bars. This can be overridden by <a
-	 * href="#plotOptions.errorbar.stemColor">stemColor</a> and <a
-	 * href="#plotOptions.errorbar.whiskerColor">whiskerColor</a> individually.
-	 * <p>
-	 * Defaults to: #000000
-	 */
-	public void setColor(Color color) {
-		this.color = color;
 	}
 
 	/**
