+++ model/src/main/java/com/vaadin/addon/charts/model/PlotOptionsScatter.java
@@ -1,8 +1,8 @@
 package com.vaadin.addon.charts.model;
 
-import com.vaadin.addon.charts.model.style.Color;
 import java.util.ArrayList;
 import java.util.Arrays;
+import com.vaadin.addon.charts.model.style.Color;
 import java.util.Date;
 import java.time.Instant;
 import com.vaadin.addon.charts.util.Util;
@@ -12,7 +12,6 @@ public class PlotOptionsScatter extends PointOptions {
 	private Boolean animation;
 	private Number animationLimit;
 	private String className;
-	private Color color;
 	private Number colorIndex;
 	private Number cropThreshold;
 	private Cursor cursor;
@@ -148,37 +147,6 @@ public class PlotOptionsScatter extends PointOptions {
 	 */
 	public void setClassName(String className) {
 		this.className = className;
-	}
-
-	/**
-	 * @see #setColor(Color)
-	 */
-	public Color getColor() {
-		return color;
-	}
-
-	/**
-	 * <p>
-	 * The main color or the series. In line type series it applies to the line
-	 * and the point markers unless otherwise specified. In bar type series it
-	 * applies to the bars unless a color is specified per point. The default
-	 * value is pulled from the <code>options.colors</code> array.
-	 * </p>
-	 * 
-	 * <p>
-	 * In <a href=
-	 * "http://www.highcharts.com/docs/chart-design-and-style/style-by-css"
-	 * >styled mode</a>, the color can be defined by the <a
-	 * href="#plotOptions.series.colorIndex">colorIndex</a> option. Also, the
-	 * series color can be set with the <code>.highcharts-series</code>,
-	 * <code>.highcharts-color-{n}</code>,
-	 * <code>.highcharts-{type}-series</code> or
-	 * <code>.highcharts-series-{n}</code> class, or individual classes given by
-	 * the <code>className</code> option.
-	 * </p>
-	 */
-	public void setColor(Color color) {
-		this.color = color;
 	}
 
 	/**
