+++ model/src/main/java/com/vaadin/addon/charts/model/PlotOptionsWaterfall.java
@@ -18,7 +18,6 @@ public class PlotOptionsWaterfall extends ColumnOptions {
 	private Number borderRadius;
 	private Number borderWidth;
 	private String className;
-	private Color color;
 	private Boolean colorByPoint;
 	private Number colorIndex;
 	private ArrayList<Color> colors;
@@ -223,37 +222,6 @@ public class PlotOptionsWaterfall extends ColumnOptions {
 	 */
 	public void setClassName(String className) {
 		this.className = className;
-	}
-
-	/**
-	 * @see #setColor(Color)
-	 */
-	public Color getColor() {
-		return color;
-	}
-
-	/**
-	 * <p>
-	 * The main color or the series. In line type series it applies to the line
-	 * and the point markers unless otherwise specified. In bar type series it
-	 * applies to the bars unless a color is specified per point. The default
-	 * value is pulled from the <code>options.colors</code> array.
-	 * </p>
-	 * 
-	 * <p>
-	 * In <a href=
-	 * "http://www.highcharts.com/docs/chart-design-and-style/style-by-css"
-	 * >styled mode</a>, the color can be defined by the <a
-	 * href="#plotOptions.series.colorIndex">colorIndex</a> option. Also, the
-	 * series color can be set with the <code>.highcharts-series</code>,
-	 * <code>.highcharts-color-{n}</code>,
-	 * <code>.highcharts-{type}-series</code> or
-	 * <code>.highcharts-series-{n}</code> class, or individual classes given by
-	 * the <code>className</code> option.
-	 * </p>
-	 */
-	public void setColor(Color color) {
-		this.color = color;
 	}
 
 	/**
