+++ model/src/main/java/com/vaadin/addon/charts/model/PointOptions.java
@@ -52,19 +52,6 @@ public abstract class PointOptions extends AbstractPlotOptions {
     public abstract void setAnimation(Boolean animation);
 
     /**
-     * @see #setColor(Color)
-     */
-    public abstract Color getColor();
-
-    /**
-     * The main color or the series. In line type series it applies to the line
-     * and the point markers unless otherwise specified. In bar type series it
-     * applies to the bars unless a color is specified per point. The default
-     * value is pulled from the <code>options.colors</code> array.
-     */
-    public abstract void setColor(Color color);
-
-    /**
      * @see #setCropThreshold(Number)
      */
     public abstract Number getCropThreshold();
